refactor(inventory): Move slot interactions to inventory class
